Renamed files to align with the documentations I made
